Adds "notifications" and "feed-items" producers
