+++ packages/server/src/async/producers/feed-items/index.js
@@ -1,0 +1,5 @@
+const { MemberJoinedToCommunity } = require('./member-joined-to-community');
+
+module.exports = {
+  MemberJoinedToCommunity,
+};
