+++ packages/server/src/async/producers/feed-items/producer-feed-items.js
@@ -1,0 +1,13 @@
+const { Producer } = require('../producer');
+
+class ProducerFeedItems extends Producer {
+
+  constructor(options = {}) {
+    super(options);
+  }
+
+}
+
+module.exports = {
+  ProducerFeedItems,
+};
