+++ packages/server/src/async/producers/index.js
@@ -1,11 +1,7 @@
-const { MeasurableStateChanged } = require('./measurable-state-changed');
-const { MeasurableStateResolved } = require('./measurable-state-resolved');
-const { MemberAddedToCommunity } = require('./member-added-to-community');
-const { MemberInvitedToCommunity } = require('./member-invited-to-community');
+const notifications = require('./notifications');
+const feedItems = require('./feed-items');
 
 module.exports = {
-  MeasurableStateChanged,
-  MeasurableStateResolved,
-  MemberAddedToCommunity,
-  MemberInvitedToCommunity,
+  ...notifications,
+  ...feedItems,
 };
