+++ packages/server/src/async/producers/notifications/index.js
@@ -1,0 +1,11 @@
+const { MeasurableStateChanged } = require('./measurable-state-changed');
+const { MeasurableStateResolved } = require('./measurable-state-resolved');
+const { MemberAddedToCommunity } = require('./member-added-to-community');
+const { MemberInvitedToCommunity } = require('./member-invited-to-community');
+
+module.exports = {
+  MeasurableStateChanged,
+  MeasurableStateResolved,
+  MemberAddedToCommunity,
+  MemberInvitedToCommunity,
+};
