+++ packages/server/src/async/producers/notifications/measurable-state-changed.js
@@ -1,7 +1,7 @@
 const assert = require('assert');
 const _ = require('lodash');
 
-const { Producer } = require('./producer');
+const { Producer } = require('./../producer');
 const { MeasurableState } = require('./measurable-state');
 
 class MeasurableStateChanged extends MeasurableState {
@@ -29,13 +29,14 @@ class MeasurableStateChanged extends MeasurableState {
       );
       const notification = await this._queueEmail(replacements);
       await this._assignAgentToNotification(creator, notification);
+
+      await this._commit();
+      return true;
     } catch (e) {
       await this._rollback();
       console.log(`stateChanged`, e.message, e);
       return false;
     }
-    await this._commit();
-    return true;
   }
 }
