+++ packages/server/src/async/producers/notifications/measurable-state-resolved.js
@@ -1,7 +1,7 @@
 const assert = require('assert');
 const _ = require('lodash');
 
-const { Producer } = require('./producer');
+const { Producer } = require('./../producer');
 const { MeasurableState } = require('./measurable-state');
 
 class MeasurableStateResolved extends MeasurableState {
@@ -40,13 +40,13 @@ class MeasurableStateResolved extends MeasurableState {
         await this._assignAgentToNotification(agent, notification);
       }
 
+      await this._commit();
+      return true;
     } catch (e) {
       await this._rollback();
       console.log(`stateResolved`, e.message, e);
       return false;
     }
-    await this._commit();
-    return true;
   }
 }
