+++ packages/server/src/async/producers/notifications/member-added-to-community.js
@@ -1,11 +1,13 @@
-const { Producer } = require('./producer');
+const { Producer } = require('./../producer');
 const { MemberToCommunity } = require('./member-to-community');
 
 class MemberAddedToCommunity extends MemberToCommunity {
+
   constructor(channelMembership = {}) {
     super(channelMembership);
     this.templateName = Producer.TEMPLATE_NAME.MEMBER_ADDED_TO_COMMUNITY;
   }
+
 }
 
 module.exports = {
