+++ packages/server/src/async/producers/producer.js
@@ -7,11 +7,8 @@ const { TEMPLATE_NAME } = require('../../models/enums/template-name');
 const { NOTIFICATION_TYPE } = require('../../models/enums/notification-type');
 const { EmailEnvelope } = require('../../models/classes/notifications');
 
-/**
- * Here "producer" class is a parent class of all another "producers",
- * for instance "MeasurableStateProducer", and another ones.
- */
 class Producer {
+
   constructor(options = {}) {
     assert(_.isObject(options), 'Options is not an object');
 
@@ -65,70 +62,6 @@ class Producer {
   async _getOptions() {
     const transaction = await this._getTransaction();
     return { transaction };
-  }
-
-  /**
-   * @param {object} replacements
-   * @return {Promise<*>}
-   * @protected
-   */
-  async _queueEmail(replacements) {
-    const template = await this._getTemplate();
-    assert(!!_.get(template, 'id'), 'Template ID is required');
-    assert(
-      !!_.get(template, 'envelopeTemplate'),
-      'Envelope Template ID is required',
-    );
-
-    const emailEnvelope = new Producer.EmailEnvelope(template.envelopeTemplate);
-    const emailEnvelope$ = emailEnvelope.mutate(replacements);
-    return await this._createEmailNotification(emailEnvelope$);
-  }
-
-  /**
-   * @param {object} agent
-   * @param {object} notification
-   * @return {Promise<*>}
-   * @protected
-   */
-  async _assignAgentToNotification(agent, notification) {
-    assert(!!agent.id, 'Agent ID is required');
-    assert(!!notification.id, 'Notification ID is required');
-
-    const data = { agentId: agent.id, notificationId: notification.id };
-    const options = await this._getOptions();
-    return await Producer.data.agentNotifications.createOne(
-      data,
-      options,
-    );
-  }
-
-  /**
-   * @param {Producer.EmailEnvelope} emailEnvelope
-   * @return {Promise<*>}
-   * @protected
-   */
-  async _createEmailNotification(emailEnvelope) {
-    return this._createNotification(emailEnvelope);
-  }
-
-  /**
-   * @param {EmailEnvelope} envelope
-   * @param {string} type
-   * @return {Promise<*>}
-   * @protected
-   */
-  async _createNotification(
-    envelope = new Producer.EmailEnvelope(),
-    type = Producer.NOTIFICATION_TYPE.EMAIL,
-  ) {
-    assert(
-      envelope instanceof Producer.EmailEnvelope,
-      'Envelope is not EmailEnvelope'
-    );
-    const data = { type, envelope: envelope };
-    const options = await this._getOptions();
-    return Producer.data.notifications.createOne(data, options);
   }
 }
